Prevent boxing on value stack but replacing implementation with long array

+++ chasm/src/commonMain/kotlin/io/github/charlietap/chasm/embedding/Global.kt
@@ -4,6 +4,7 @@ import io.github.charlietap.chasm.ast.type.GlobalType
 import io.github.charlietap.chasm.embedding.shapes.Global
 import io.github.charlietap.chasm.embedding.shapes.Store
 import io.github.charlietap.chasm.executor.instantiator.allocation.global.GlobalAllocator
+import io.github.charlietap.chasm.executor.runtime.ext.toLong
 import io.github.charlietap.chasm.executor.runtime.instance.ExternalValue
 import io.github.charlietap.chasm.executor.runtime.value.ExecutionValue
 
@@ -24,5 +25,5 @@ internal fun global(
     value: ExecutionValue,
     allocator: GlobalAllocator,
 ): Global {
-    return Global(ExternalValue.Global(allocator(store.store, type, value)))
+    return Global(ExternalValue.Global(allocator(store.store, type, value.toLong())))
 }
