+++ chasm/src/commonMain/kotlin/io/github/charlietap/chasm/embedding/table/WriteTable.kt
@@ -10,6 +10,7 @@ import io.github.charlietap.chasm.embedding.shapes.ChasmResult.Error
 import io.github.charlietap.chasm.embedding.shapes.ChasmResult.Success
 import io.github.charlietap.chasm.embedding.shapes.Store
 import io.github.charlietap.chasm.embedding.shapes.Table
+import io.github.charlietap.chasm.executor.runtime.encoder.toLong
 import io.github.charlietap.chasm.executor.runtime.error.ModuleTrapError
 import io.github.charlietap.chasm.executor.runtime.ext.table
 import io.github.charlietap.chasm.executor.runtime.value.ReferenceValue
@@ -36,5 +37,5 @@ internal fun internalWriteTable(
     value: ReferenceValue,
 ): Result<Unit, ModuleTrapError> = binding {
     val table = store.store.table(table.reference.address)
-    table.elements[elementIndex] = value
+    table.elements[elementIndex] = value.toLong()
 }
