+++ executor/invoker/src/commonMain/kotlin/io/github/charlietap/chasm/executor/invoker/instruction/reference/RefFuncExecutor.kt
@@ -1,6 +1,7 @@
 package io.github.charlietap.chasm.executor.invoker.instruction.reference
 
 import io.github.charlietap.chasm.executor.invoker.ext.functionAddress
+import io.github.charlietap.chasm.executor.runtime.encoder.toLong
 import io.github.charlietap.chasm.executor.runtime.execution.ExecutionContext
 import io.github.charlietap.chasm.executor.runtime.instruction.ReferenceInstruction
 import io.github.charlietap.chasm.executor.runtime.value.ReferenceValue
@@ -15,5 +16,5 @@ internal inline fun RefFuncExecutor(
     val functionAddress = frame.instance
         .functionAddress(instruction.funcIdx)
 
-    stack.push(ReferenceValue.Function(functionAddress))
+    stack.push(ReferenceValue.Function(functionAddress).toLong())
 }
