+++ test/fixture/executor/runtime/src/commonMain/kotlin/io/github/charlietap/chasm/fixture/executor/runtime/stack/ValueStack.kt
@@ -1,5 +1,6 @@
 package io.github.charlietap.chasm.fixture.executor.runtime.stack
 
+import io.github.charlietap.chasm.executor.runtime.ext.toLong
 import io.github.charlietap.chasm.executor.runtime.stack.ValueStack
 import io.github.charlietap.chasm.executor.runtime.value.ExecutionValue
 
@@ -7,6 +8,6 @@ fun vstack(
     values: List<ExecutionValue> = emptyList(),
 ) = ValueStack().apply {
     values.forEach { value ->
-        push(value)
+        push(value.toLong())
     }
 }
